change field data topup table migration

+++ db/migrations/000006_create_topup_table.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE "topup" (
     "id" varchar(100) PRIMARY KEY,
-    "user_id" int NOT NULL,
+    "user_id" bigint NOT NULL,
     "amount" bigint NOT NULL,
     "status" int NOT NULL,
     "snap_url" varchar(255)
